create generator class
